Rename test/type-syntax.ts to test/flow.ts.
